Created StatusController,StatusService, removed whitespaces.

+++ webshop/src/main/java/elte/alkfejlbead/webshop/repository/StatusRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StatusRepository extends CrudRepository<Status, Long>{
+public interface StatusRepository extends CrudRepository<Status, Integer>{
     List<Status> findAllByOrderStatus(Status.OrderStatus orderStatus);
 
     List<Status> findAllByShippingStatus(Status.ShippingStatus shippingStatus);
